Improve watch-test experience
